Refactor: rename user reducer
